+++ packages/emails/src/templates/BookingRedirectEmailNotification.tsx
@@ -1,4 +1,4 @@
-import type { IBookingRedirect } from "../../templates/booking-redirect-notification";
+import type { IBookingRedirect } from "../../lib/utils/booking-redirect-types";
 import { BaseEmailHtml } from "../components";
 
 export const BookingRedirectEmailNotification = (
