+++ packages/emails/src/templates/VerifyEmailByCode.tsx
@@ -1,6 +1,6 @@
 import { APP_NAME, SENDER_NAME, SUPPORT_MAIL_ADDRESS } from "@calcom/lib/constants";
 
-import type { EmailVerifyCode } from "../../templates/attendee-verify-email";
+import type { EmailVerifyCode } from "../../lib/utils/email-types";
 import { BaseEmailHtml } from "../components";
 
 export const VerifyEmailByCode = (
