+++ packages/emails/templates/attendee-cancelled-seat-email.ts
@@ -1,6 +1,6 @@
 import { getReplyToHeader } from "@calcom/lib/getReplyToHeader";
 
-import { renderEmail } from "../";
+import renderEmail from "../src/renderEmail";
 import AttendeeScheduledEmail from "./attendee-scheduled-email";
 
 export default class AttendeeCancelledSeatEmail extends AttendeeScheduledEmail {
