+++ packages/emails/templates/attendee-declined-email.ts
@@ -1,7 +1,7 @@
 import { getReplyToHeader } from "@calcom/lib/getReplyToHeader";
 import type { CalendarEvent, Person } from "@calcom/types/Calendar";
 
-import { renderEmail } from "../";
+import renderEmail from "../src/renderEmail";
 import AttendeeScheduledEmail from "./attendee-scheduled-email";
 
 export default class AttendeeDeclinedEmail extends AttendeeScheduledEmail {
