+++ packages/emails/templates/attendee-verify-email.ts
@@ -1,19 +1,10 @@
-import type { TFunction } from "i18next";
-
 import { APP_NAME, COMPANY_NAME, EMAIL_FROM_NAME } from "@calcom/lib/constants";
 
-import { renderEmail } from "../";
+import type { EmailVerifyCode } from "../lib/utils/email-types";
+import renderEmail from "../src/renderEmail";
 import BaseEmail from "./_base-email";
 
-export type EmailVerifyCode = {
-  language: TFunction;
-  user: {
-    name?: string | null;
-    email: string;
-  };
-  verificationEmailCode: string;
-  isVerifyingEmail?: boolean;
-};
+export type { EmailVerifyCode } from "../lib/utils/email-types";
 
 export default class AttendeeVerifyEmail extends BaseEmail {
   verifyAccountInput: EmailVerifyCode;
