+++ packages/emails/templates/credit-balance-limit-reached-email.ts
@@ -2,7 +2,7 @@ import type { TFunction } from "i18next";
 
 import { EMAIL_FROM_NAME } from "@calcom/lib/constants";
 
-import { renderEmail } from "..";
+import renderEmail from "../src/renderEmail";
 import BaseEmail from "./_base-email";
 
 export default class CreditBalanceLimitReachedEmail extends BaseEmail {
