+++ packages/emails/templates/feedback-email.ts
@@ -1,6 +1,6 @@
 import { EMAIL_FROM_NAME } from "@calcom/lib/constants";
 
-import { renderEmail } from "../";
+import renderEmail from "../src/renderEmail";
 import BaseEmail from "./_base-email";
 
 export interface Feedback {
