+++ packages/emails/templates/forgot-password-email.ts
@@ -2,7 +2,7 @@ import type { TFunction } from "i18next";
 
 import { APP_NAME, EMAIL_FROM_NAME } from "@calcom/lib/constants";
 
-import { renderEmail } from "../";
+import renderEmail from "../src/renderEmail";
 import BaseEmail from "./_base-email";
 
 export type PasswordReset = {
