+++ packages/emails/templates/monthly-digest-email.ts
@@ -1,6 +1,6 @@
 import { APP_NAME, EMAIL_FROM_NAME } from "@calcom/lib/constants";
 
-import { renderEmail } from "../";
+import renderEmail from "../src/renderEmail";
 import type { MonthlyDigestEmailData } from "../src/templates/MonthlyDigestEmail";
 import BaseEmail from "./_base-email";
