+++ packages/emails/templates/organizer-add-guests-email.ts
@@ -1,8 +1,8 @@
 import { APP_NAME } from "@calcom/lib/constants";
 import { getReplyToHeader } from "@calcom/lib/getReplyToHeader";
 
-import { renderEmail } from "../";
 import generateIcsFile, { GenerateIcsRole } from "../lib/generateIcsFile";
+import renderEmail from "../src/renderEmail";
 import OrganizerScheduledEmail from "./organizer-scheduled-email";
 
 export default class OrganizerAddGuestsEmail extends OrganizerScheduledEmail {
