+++ packages/emails/templates/organizer-attendee-cancelled-seat-email.ts
@@ -1,6 +1,6 @@
 import { EMAIL_FROM_NAME } from "@calcom/lib/constants";
 
-import { renderEmail } from "../";
+import renderEmail from "../src/renderEmail";
 import OrganizerScheduledEmail from "./organizer-scheduled-email";
 
 export default class OrganizerCancelledEmail extends OrganizerScheduledEmail {
